+++ HotelBack/src/main/java/com/example/hotelback/repositories/CabinRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CabinRepository extends JpaRepository<Cabin, Integer> {
+public interface CabinRepository extends JpaRepository<Cabin,Long> {
 
-    Cabin findByIdcabin(Integer idcabin);
+    Cabin findByIdcabin(Long idcabin);
 }
